Working game with the new structure. #43
