+++ reflexy/logic/ann/player_logic.py
@@ -1,0 +1,2 @@
+def Brain():
+    pass
