+++ reflexy/models/weapons/ray.py
@@ -1,12 +1,13 @@
 import math
 import pygame
-from reflexy.helpers import draw_box, get_image_path, get_surface
+from reflexy.helpers.general import draw_box, get_surface
 from reflexy.constants import (
     RAY_WIDTH,
     RAY_HEIGHT,
     RAY_ORIGIN_X,
     RAY_ANIMATION_TIME,
 )
+from typing import Sequence
 
 
 class Ray(pygame.sprite.Sprite):
